+++ include/beast/http.hpp
@@ -21,6 +21,7 @@
 #include <beast/http/message_parser.hpp>
 #include <beast/http/read.hpp>
 #include <beast/http/rfc7230.hpp>
+#include <beast/http/serializer.hpp>
 #include <beast/http/string_body.hpp>
 #include <beast/http/write.hpp>
